fix: Spring security 보완

+++ src/main/java/org/example/wantedpreonboardingchallengebackend20/product/repository/ProductRepository.java
@@ -11,5 +11,5 @@ import java.util.List;
 public interface ProductRepository extends JpaRepository<Product, Long> {
     Product save(ProductRequest product);
     List<Product> findAll();
-    Product findById(int id);
+    Product findById(long id);
 }
