+++ src/main/java/org/example/wantedpreonboardingchallengebackend20/product/service/ProductService.java
@@ -1,6 +1,7 @@
 package org.example.wantedpreonboardingchallengebackend20.product.service;
 
 import lombok.extern.slf4j.Slf4j;
+import org.example.wantedpreonboardingchallengebackend20.order.service.OrderService;
 import org.example.wantedpreonboardingchallengebackend20.product.entity.Product;
 import org.example.wantedpreonboardingchallengebackend20.product.repository.ProductRepository;
 import org.springframework.stereotype.Service;
@@ -26,7 +27,7 @@ public class ProductService {
         return productRepository.save(product);
     }
 
-    public Product getProductById(int productId) {
+    public Product getProductById(long productId) {
         return productRepository.findById(productId);
     }
 }
